refactor(status): extract benchmark core projection
